Removes the IPC for launching the app from the updater.
First, the updater don't have to push information to the app process (as far as I can see).
Second, this causes major trouble with multiple updates with "update chains" (multiple restarting updates without existing the application between the updates). This comes because the first update process still blocks the named pipe making it impossible for the second update to start the cold update process.